Moved inst_update_partition_auto to lib.
