+++ src/Makefile.am
@@ -30,6 +30,10 @@ ybin_SCRIPTS = \
 desktop_DATA = \
   desktop/update.desktop
 
-EXTRA_DIST = $(module_DATA) $(client_DATA) $(ynclude_DATA) $(ybin_SCRIPTS) $(desktop_DATA)
+ylibclientdir = "${yast2dir}/lib/update/clients"
+ylibclient_DATA = \
+  lib/update/clients/inst_update_partition_auto.rb
+
+EXTRA_DIST = $(module_DATA) $(client_DATA) $(ynclude_DATA) $(ybin_SCRIPTS) $(desktop_DATA) $(ylibclient_DATA)
 
 include $(top_srcdir)/Makefile.am.common
